+++ src/scratchgrad/utils/__init__.py
@@ -5,6 +5,7 @@ from scratchgrad.utils.validation import (
     check_array,
     check_is_fitted,
     check_random_state,
+    check_sample_weight,
     check_X_y,
 )
 
@@ -13,6 +14,7 @@ __all__ = [
     "check_array",
     "check_is_fitted",
     "check_random_state",
+    "check_sample_weight",
     "logsumexp",
     "sigmoid",
     "softmax",